EmotionCaptureDone

+++ AIFinal/AIFinal/App_Code/Emotion.cs
@@ -10,8 +10,8 @@ public class Emotion
 {
     public string Mp3ParentName;
     public string emotionDetected { get; set; }
-    public int intensity { get; set; }
-    public int delta { get; set; }
+    public double intensity { get; set; }
+    public double delta { get; set; }
     public string[] ArrayWords;
 
     public Emotion()
